# static/swagger.yaml
openapi: 3.0.1
info:
  title: MEDIATHEQUE
  version: 1.0.5
  description: API de gestion des abonnés et des emprunts pour une médiathèque.
servers:
  - url: "http://localhost:5000"
    description: Serveur de développement
...
paths:
  /abonnee/insert_abonnee:
    post:
      tags:
      - Gestion Abonne
      summary: Ajouter un nouvel abonné
      description: Cette route permet d'ajouter un nouvel abonné dans la médiathèque.
      operationId: ajouterAbonne
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  description: Nom de l'abonné
                prenom:
                  type: string
                  description: Prénom de l'abonné
                adresse:
                  type: string
                  description: Adresse de l'abonné
      responses:
        '201':
          description: Abonné ajouté avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
        '400':
          description: Champs manquants ou invalides
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...